test: move crash diagnostics coverage to Swift Testing
