+++ cmuxTests/SessionPersistenceTests.swift
@@ -6001,94 +6001,6 @@ extension SessionPersistenceTests {
         XCTAssertNil(cleanScanBinding)
     }
 
-    @MainActor
-    func testAppDelegateSessionSnapshotDropsCrashDiagnosticWindow() throws {
-        let previousAppDelegate = AppDelegate.shared
-        let app = AppDelegate()
-        AppDelegate.shared = app
-        defer {
-            AppDelegate.shared = previousAppDelegate
-        }
-
-        let projectDirectory = "/tmp/cmux-project"
-        let projectManager = TabManager(
-            initialWorkingDirectory: projectDirectory,
-            autoWelcomeIfNeeded: false
-        )
-        let projectWindowId = app.registerMainWindowContextForTesting(tabManager: projectManager)
-        defer {
-            app.unregisterMainWindowContextForTesting(windowId: projectWindowId)
-        }
-
-        let crashDirectory = FileManager.default.homeDirectoryForCurrentUser
-            .appendingPathComponent(".local/state/cmux/crash", isDirectory: true)
-            .path
-        let crashManager = TabManager(
-            initialWorkingDirectory: crashDirectory,
-            autoWelcomeIfNeeded: false
-        )
-        let crashWindowId = app.registerMainWindowContextForTesting(tabManager: crashManager)
-        defer {
-            app.unregisterMainWindowContextForTesting(windowId: crashWindowId)
-        }
-
-        let snapshot = try XCTUnwrap(app.debugBuildSessionSnapshotForTesting(includeScrollback: false))
-        let restoredDirectories = snapshot.windows.flatMap { window in
-            window.tabManager.workspaces.map(\.currentDirectory)
-        }
-
-        XCTAssertEqual(snapshot.windows.count, 1)
-        XCTAssertEqual(restoredDirectories, [projectDirectory])
-    }
-
-    func testSessionSnapshotDropsEmptyCrashDiagnosticWorkspace() {
-        let projectDirectory = "/tmp/cmux-project"
-        let crashDirectory = FileManager.default.homeDirectoryForCurrentUser
-            .appendingPathComponent(".local/state/cmux/crash", isDirectory: true)
-            .path
-        let snapshot = AppSessionSnapshot(
-            version: SessionSnapshotSchema.currentVersion,
-            createdAt: 10,
-            windows: [
-                SessionWindowSnapshot(
-                    frame: nil,
-                    display: nil,
-                    tabManager: SessionTabManagerSnapshot(
-                        selectedWorkspaceIndex: 0,
-                        workspaces: [
-                            emptyWorkspaceSnapshot(currentDirectory: crashDirectory),
-                            emptyWorkspaceSnapshot(currentDirectory: projectDirectory),
-                        ]
-                    ),
-                    sidebar: SessionSidebarSnapshot(isVisible: true, selection: .tabs, width: nil)
-                ),
-            ]
-        )
-
-        let pruned = SessionPersistencePolicy.pruningCmuxCrashDiagnosticWindows(from: snapshot)
-
-        XCTAssertTrue(pruned.removedAny)
-        XCTAssertEqual(pruned.snapshot?.windows.first?.tabManager.workspaces.map(\.currentDirectory), [projectDirectory])
-        XCTAssertEqual(pruned.snapshot?.windows.first?.tabManager.selectedWorkspaceIndex, 0)
-    }
-
-    private func emptyWorkspaceSnapshot(currentDirectory: String) -> SessionWorkspaceSnapshot {
-        SessionWorkspaceSnapshot(
-            processTitle: "Terminal",
-            customTitle: nil,
-            customColor: nil,
-            isPinned: false,
-            currentDirectory: currentDirectory,
-            focusedPanelId: nil,
-            layout: .pane(SessionPaneLayoutSnapshot(panelIds: [], selectedPanelId: nil)),
-            panels: [],
-            statusEntries: [],
-            logEntries: [],
-            progress: nil,
-            gitBranch: nil
-        )
-    }
-
     func testTmuxProcessDetectedResumeBindingPreservesSocketFlags() throws {
         let binding = try XCTUnwrap(
             SurfaceResumeBindingIndex.tmuxResumeBindingForTesting(
